requirements.txt
aiohttp>=3.10
PyYAML>=6.0
